Bas' comment - make UserIdCtxKey private to authentication package.
